requirements.txt
pandas==1.0.5
scenera.node==0.1.12
pymongo==3.11.0
Flask_Cors==3.0.9
numpy==1.19.3
Flask==1.1.2
opencv_python==4.5.2.54
requests==2.21.0
msrest==0.6.19
azure_storage==0.37.0
Pillow==8.4.0
